+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+- Handle HTTP/2 half-closed connections gracefully. (#777)
 - Add support for Python 3.12. (#807)
 
 ## 0.18.0 (September 8th, 2023)